added start page and made a select dropdown
