+++ code/src/components/SecondQuestion.js
@@ -9,12 +9,17 @@ const SecondQuestion = ({
     <form>
       <h2>second question</h2>
       <label htmlFor="interstInput">What is your interest?</label>
-      <input
+      <select
         id="interestInput"
         type="text"
         value={interestInput}
         onChange={onInterestInputChange}
-      />
+      >
+        <option value>Select what you are interested in</option>
+        <option value="buyer">Buyer</option>
+        <option value="seller">Seller</option>
+        <option value="buyer-and-seller">Buyer and seller</option>
+      </select>
       <button onClick={onStepChange}>Next question</button>
     </form>
   );
