+++ code/src/components/Summary.js
@@ -2,13 +2,12 @@ import React from "react";
 
 const Summary = ({ nameInput, interestInput, ageInput, onStepChange }) => {
   return (
-    <div>
+    <section>
       <h2>Summary</h2>
       <p>Your name is: {nameInput}</p>
       <p>You want to be: {interestInput} </p>
       <p>Your age is between: {ageInput} </p>
-      <button onClick={onStepChange}>Submit</button>
-    </div>
+    </section>
   );
 };
